refactor(front): move tags and webhook_source DTO types out of lib/api into types/api
Twelfth batch of the lib/api DTO migration: the last clean internal files.
Transport DTO types and zod schemas move to types/api/*, leaving business logic
and domain types in place.

- tags: moved wholesale (GetTagsResponseBody, CreateTagResponseBody).
- webhook_source: trimmed — moved 10 transport DTOs/schemas; kept all business
  logic (deleteWebhookSource, listWebhookSourcesWithCounts, ...) and domain types
  (WebhookSourceWithCounts, WebhookSourceAdminDetails); dropped unused imports.
- 7 consumers updated; 1 mixed-import route split.

This completes the clean internal DTO set. Remaining: public-v1 DTOs (SDK track),
io-ts configuration/triggers.ts, the credentials.ts naming collision, and the
gotcha files.
