+++ front-api/routes/w/[wId]/webhook_sources/index.ts
@@ -1,17 +1,17 @@
 import { getWebhookSourcesUsage } from "@app/lib/api/agent_triggers";
 import config from "@app/lib/api/config";
 import { WEBHOOK_SERVICES } from "@app/lib/api/triggers/built-in-webhooks/services";
-import {
-  deleteWebhookSource,
-  type GetWebhookSourcesResponseBody,
-  type PostWebhookSourcesResponseBody,
-} from "@app/lib/api/webhook_source";
+import { deleteWebhookSource } from "@app/lib/api/webhook_source";
 import { SpaceResource } from "@app/lib/resources/space_resource";
 import { generateSecureSecret } from "@app/lib/resources/string_ids_server";
 import { WebhookSourceResource } from "@app/lib/resources/webhook_source_resource";
 import { WebhookSourcesViewResource } from "@app/lib/resources/webhook_sources_view_resource";
 import { buildWebhookUrl } from "@app/lib/webhook_source";
 import logger from "@app/logger/logger";
+import type {
+  GetWebhookSourcesResponseBody,
+  PostWebhookSourcesResponseBody,
+} from "@app/types/api/webhook_source";
 import type { ModelId } from "@app/types/shared/model_id";
 import { WebhookSourcesSchema } from "@app/types/triggers/webhooks";
 import { workspaceApp } from "@front-api/middlewares/ctx";
